Extract canvas hit testing helpers
